Week 01, Assignment _1
Create a webpage that displays information about a conference. Use h1 tags to define the conference title, p tags to add information about the conference, and ul tags to list the conference schedule. Use a tags to link to the conference registration page and img tags to display the conference logo.